Initial partial work for zipper merkle trees
Regular merkle trees were far too slow when used with large data set
when diffing by virtue of needing an O(log N) operation for each lookup
(changing things to something like O(n log n) for total diff instead of
O(log n) on average, yielding terrible run times. The diffing from the
old local algorithm was however still very fast.

In an attempt to speed things up, I thought of implementing the same
merkle tree, but using zippers so that search state might be cached and
reused to bring the runtime closer to the one we'd have in a local
diffing.

This commit contains initial work to build and navigate the tree.
Diffing functions are still not implemented.