Improve diff time with record affinity
By hacking around the internal record structure so that a leaf's hash
and an inner node's children hash happen in the same position in a
tuple, diffing can be sped up by 5% up to 20% by removing the need to
compare the node type to get its hash value.

The impact is significant as trees become bigger and diffsets larger.
This does mean, however that versions of a tree are no longer compatible
with newer code, even though the interfaces and usage remains the same.

Live code updating wouldn't be possible across versions with this change
in.